Schema editor: Add virtualisation
