--- packages/web-console/src/scenes/Import/ImportCSVFiles/table-schema-columns.tsx
@@ -1,219 +1,0 @@
-import React, { useEffect, useState } from "react"
-import styled from "styled-components"
-import { useFieldArray, useFormContext } from "react-hook-form"
-import { Box } from "../../../components/Box"
-import { Form } from "../../../components/Form"
-import { IconWithTooltip } from "../../../components"
-import { Button } from "@questdb/react-components"
-import { AddCircle, Close } from "styled-icons/remix-line"
-import { SortDown } from "styled-icons/boxicons-regular"
-import { Text } from "../../../components/Text"
-import { Drawer } from "../../../components/Drawer"
-import { DEFAULT_TIMESTAMP_FORMAT } from "./const"
-import { SchemaColumn } from "./types"
-
-const supportedColumnTypes: { label: string; value: string }[] = [
-  { label: "AUTO", value: "" },
-  { label: "BOOLEAN", value: "BOOLEAN" },
-  { label: "BYTE", value: "BYTE" },
-  { label: "DOUBLE", value: "DOUBLE" },
-  { label: "DATE", value: "DATE" },
-  { label: "FLOAT", value: "FLOAT" },
-  { label: "GEOHASH", value: "GEOHASH" },
-  { label: "INT", value: "INT" },
-  { label: "LONG", value: "LONG" },
-  { label: "SHORT", value: "SHORT" },
-  { label: "CHAR", value: "CHAR" },
-  { label: "STRING", value: "STRING" },
-  { label: "SYMBOL", value: "SYMBOL" },
-  { label: "TIMESTAMP", value: "TIMESTAMP" },
-]
-
-const Row = styled(Box).attrs({
-  flexDirection: "column",
-  align: "flex-start",
-  gap: "2rem",
-})`
-  width: 100%;
-`
-
-const Columns = styled(Box).attrs({
-  align: "flex-start",
-  justifyContent: "space-around",
-  gap: "1rem",
-})`
-  width: 100%;
-
-  > button {
-    padding-left: 0.5rem;
-    padding-right: 0.5rem;
-  }
-`
-
-const AddBox = styled(Box).attrs({
-  align: "center",
-  justifyContent: "center",
-})`
-  margin: auto;
-`
-
-export const TableSchemaColumns = ({ schema }: { schema: SchemaColumn[] }) => {
-  const { setValue, watch } = useFormContext()
-  const { fields, append, remove } = useFieldArray({
-    name: "schemaColumns",
-  })
-  const [columns, setColumns] = useState<SchemaColumn[]>(schema)
-
-  const watchTimestamp = watch("timestamp")
-
-  useEffect(() => {
-    setValue("schemaColumns", schema)
-  }, [schema])
-
-  return (
-    <>
-      {fields.map((field, index) => (
-        <Drawer.GroupItem direction="column" key={field.id}>
-          <Row
-            gap="2rem"
-            flexDirection="column"
-            align="flex-start"
-            key={field.id}
-          >
-            <Columns>
-              <Text color="foreground">{index + 1}</Text>
-              <Box gap="1rem" flexDirection="column">
-                <Box gap="1rem" align="flex-end">
-                  <Form.Item name={`schemaColumns.${index}.name`} label="Name">
-                    <Form.Input
-                      name={`schemaColumns.${index}.name`}
-                      onChange={(e) => {
-                        const cols = [...columns]
-                        cols[index].name = e.target.value
-                        setColumns(cols)
-                      }}
-                    />
-                  </Form.Item>
-                  <Form.Item name={`schemaColumns.${index}.type`} label="Type">
-                    <Form.Select
-                      name={`schemaColumns.${index}.type`}
-                      options={supportedColumnTypes}
-                      onChange={(e) => {
-                        const cols = [...columns]
-                        cols[index].type = e.target.value
-                        setColumns(cols)
-                      }}
-                    />
-                  </Form.Item>
-
-                  <IconWithTooltip
-                    icon={
-                      <Button
-                        disabled={
-                          columns[index] && columns[index].type !== "TIMESTAMP"
-                        }
-                        skin={
-                          watchTimestamp !== "" &&
-                          columns[index] &&
-                          columns[index].name !== "" &&
-                          watchTimestamp === columns[index].name
-                            ? "success"
-                            : "transparent"
-                        }
-                        onClick={() => {
-                          setValue(
-                            "timestamp",
-                            watchTimestamp === columns[index].name
-                              ? undefined
-                              : columns[index].name,
-                          )
-                        }}
-                        type="button"
-                      >
-                        <SortDown size="18px" />
-                      </Button>
-                    }
-                    tooltip="Set as designated timestamp"
-                    placement="bottom"
-                  />
-
-                  <Button
-                    skin="transparent"
-                    onClick={() => {
-                      remove(index)
-                      if (watchTimestamp === columns[index].name) {
-                        setValue("timestamp", undefined)
-                      }
-                      setColumns(columns.filter((_, i) => i !== index))
-                    }}
-                    type="button"
-                  >
-                    <Close size="18px" />
-                  </Button>
-                </Box>
-                {columns[index] && columns[index].type === "TIMESTAMP" && (
-                  <Form.Item
-                    name={`schemaColumns.${index}.pattern`}
-                    label="Timestamp pattern"
-                    helperText="Required when using the TIMESTAMP type"
-                  >
-                    <Form.Input
-                      name={`schemaColumns.${index}.pattern`}
-                      placeholder={DEFAULT_TIMESTAMP_FORMAT}
-                      defaultValue={DEFAULT_TIMESTAMP_FORMAT}
-                      required
-                    />
-                  </Form.Item>
-                )}
-                {columns[index] && columns[index].type === "GEOHASH" && (
-                  <Form.Item
-                    name={`schemaColumns.${index}.precision`}
-                    label="Precision"
-                    helperText={
-                      <a
-                        href="https://questdb.io/docs/concept/geohashes/#specifying-geohash-precision"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                      >
-                        Docs on QuestDB geohash precision
-                      </a>
-                    }
-                  >
-                    <Form.Input
-                      name={`schemaColumns.${index}.precision`}
-                      required
-                    />
-                  </Form.Item>
-                )}
-              </Box>
-            </Columns>
-          </Row>
-        </Drawer.GroupItem>
-      ))}
-
-      <Drawer.GroupItem direction="column">
-        <AddBox>
-          <Button
-            prefixIcon={<AddCircle size="18px" />}
-            skin="transparent"
-            onClick={() => {
-              const newColumn = {
-                name: "",
-                type: "",
-                pattern: "",
-                precision: "",
-              }
-              append(newColumn)
-              setColumns([...columns, newColumn])
-            }}
-            type="button"
-          >
-            Add column
-          </Button>
-        </AddBox>
-      </Drawer.GroupItem>
-
-      <Form.Input name="timestamp" hidden />
-    </>
-  )
-}
